+++ .gitignore
@@ -5,7 +5,7 @@ data/**/*.mat
 .idea/
 .vscode/
 __pycache__/
-src/__pycache__/
+pyad/__pycache__/
 __pycache__
 results/**/*.txt
 models/**/*.pt
