+++ notebooks/20-05-2022/test.py
@@ -4,20 +4,20 @@ import re
 
 import numpy as np
 
-from src.bootstrap import store_results
-from src.datamanager.dataset import ArrhythmiaDataset, ThyroidDataset, IDS2017Dataset, IDS2018Dataset
-from src.trainer.adversarial import ALADTrainer
-from src.trainer.density import DSEBMTrainer
-from src.trainer.one_class import DeepSVDDTrainer, EdgeMLDROCCTrainer
-from src.trainer.reconstruction import AutoEncoderTrainer, MemAETrainer, DAGMMTrainer
-from src.trainer.transformers import NeuTraLADTrainer
-from src.model.adversarial import ALAD
-from src.model.density import DSEBM
-from src.model.one_class import DeepSVDD, DROCC
-from src.model.reconstruction import AutoEncoder, DAGMM, MemAutoEncoder
-from src.model.transformers import NeuTraLAD
-from src.utils import metrics
-from src.utils.utils import ids_misclf_per_label
+from pyad.bootstrap import store_results
+from pyad.datamanager.dataset import ArrhythmiaDataset, ThyroidDataset, IDS2017Dataset, IDS2018Dataset
+from pyad.trainer.adversarial import ALADTrainer
+from pyad.trainer.density import DSEBMTrainer
+from pyad.trainer.one_class import DeepSVDDTrainer, EdgeMLDROCCTrainer
+from pyad.trainer.reconstruction import AutoEncoderTrainer, MemAETrainer, DAGMMTrainer
+from pyad.trainer.transformers import NeuTraLADTrainer
+from pyad.model.adversarial import ALAD
+from pyad.model.density import DSEBM
+from pyad.model.one_class import DeepSVDD, DROCC
+from pyad.model.reconstruction import AutoEncoder, DAGMM, MemAutoEncoder
+from pyad.model.transformers import NeuTraLAD
+from pyad.utils import metrics
+from pyad.utils.utils import ids_misclf_per_label
 from train import settings
 import pandas as pd
