+++ pyad/trainer/transformers.py
@@ -1,8 +1,8 @@
 import torch
 from torch import nn, optim
 from torch.optim.lr_scheduler import StepLR
-from src.model.transformers import NeuTraLAD
-from src.trainer.base import BaseTrainer
+from pyad.model.transformers import NeuTraLAD
+from pyad.trainer.base import BaseTrainer
 
 
 class NeuTraLADTrainer(BaseTrainer):
