+++ pyad/utils/utils.py
