+++ requirements.txt
@@ -9,3 +9,4 @@ pandas~=1.2.1
 joblib~=1.0.0
 scipy~=1.6.0
 setuptools~=56.1.0
+ray~=1.11.0
